# Improve ecs example performance