--- a/cmd/ebiten-ecs/system-draw.go
+++ b/cmd/ebiten-ecs/system-draw.go
@@ -53,14 +53,14 @@ func (s *systemDraw) Run(world *ClientWorld, screen *ebiten.Image) {
 
 	draw.Draw(s.buffer, s.buffer.Bounds(), &image.Uniform{color.Transparent}, image.Point{}, draw.Src)
 
-	components.Color.AllParallel(func(entity ecs.EntityID, color *color.RGBA) bool {
-		if color == nil {
+	components.Health.AllParallel(func(entity ecs.EntityID, h *health) bool {
+		if h == nil {
 			return true
 		}
 
 		transform := components.Transform.Get(entity)
 
-		s.buffer.SetRGBA(int(transform.x), int(transform.y), *color)
+		s.buffer.SetRGBA(int(transform.x), int(transform.y), h.color)
 		return true
 	})
 
@@ -89,7 +89,6 @@ func (s *systemDraw) Run(world *ClientWorld, screen *ebiten.Image) {
 	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Entity count %d", world.Size()))
 	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Transforms count %d", components.Transform.Len()))
 	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Healths count %d", components.Health.Len()))
-	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Colors count %d", components.Color.Len()))
 	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Cameras count %d", components.Camera.Len()))
 	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Destroys count %d", components.Destroy.Len()))
 	s.debugInfo = append(s.debugInfo, s.p.Sprintf("Pprof %d", systems.Spawn.pprofEnabled))