… cmd/ebiten-ecs/system-calc-color.go