Ideas behind change:

- replace haxmap.Map with golang's default map (gave 20 fps -> 60 fps on 100k entities boost)
- remove color subsystem by merge with with health subsystem and auto-updating colocated values (gave 20 -> 60 fps on 200k enities boost)

![image](https://github.com/user-attachments/assets/b6688d27-4b10-4f1b-9c4b-d80fb4aaebd0)