@melg8 2024-12-29
Was partially added to main source code already, so closing this.